+++ packages/gov-ui-kit/src/core/components/toggles/toggleGroup/toggleGroup.tsx
@@ -7,6 +7,11 @@ export type ToggleGroupValue<TMulti extends boolean> = TMulti extends true ? str
 export interface IToggleGroupBaseProps<TMulti extends boolean>
     extends Omit<ComponentProps<'div'>, 'value' | 'onChange' | 'defaultValue' | 'ref' | 'dir'> {
     /**
+     * Variant of the component defining the spacing between the toggle items.
+     * @default fixed
+     */
+    variant?: 'fixed' | 'space-between';
+    /**
      * Allows multiple toggles to be selected at the same time when set to true.
      */
     isMultiSelect: TMulti;
@@ -27,8 +32,14 @@ export interface IToggleGroupBaseProps<TMulti extends boolean>
 export type IToggleGroupProps = IToggleGroupBaseProps<true> | IToggleGroupBaseProps<false>;
 
 export const ToggleGroup = (props: IToggleGroupProps) => {
-    const { value, defaultValue, onChange, isMultiSelect, className, ...otherProps } = props;
-    const classes = classNames('flex flex-row flex-wrap gap-2 md:gap-3', className);
+    const { variant = 'fixed', value, defaultValue, onChange, isMultiSelect, className, ...otherProps } = props;
+
+    const classes = classNames(
+        'flex flex-row flex-wrap',
+        { 'gap-2 md:gap-3': variant === 'fixed' },
+        { 'justify-between gap-y-2': variant === 'space-between' },
+        className,
+    );
 
     if (isMultiSelect) {
         return (