+++ packages/gov-ui-kit/CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 - Update `<DaoDataListItem.Structure />` module component to support new `isExternal` property
+- Support `xs` size on `<Avatar />` core component
+- Add `variant` property to `<ToggleGroup />` core component to support space-between spacing
 
 ### Changed